Mostly for replacing boost library calls, but some of the Direct3D API calls in post-process effect code were commented out as well.

+++ framework/amorphous/3DMath/ConvexPolygonUtilities.hpp
@@ -107,7 +107,7 @@ inline void TriangulateConvexPolygon(
 
 	triangle_indices.reserve( point_indices.size() );
 	
-	for( int num_orig_indices = indices.size();
+	for( int num_orig_indices = (int)indices.size();
 		0 < num_orig_indices;
 		num_orig_indices--)
 	{
